re-fixed format specifier for amd64

+++ Validation/RecoVertex/src/PrimaryVertexAnalyzer4PU.cc
@@ -2248,7 +2248,7 @@ PrimaryVertexAnalyzer4PU::analyze(const Event& iEvent, const EventSetup& iSetup)
       if(printXBS_) {
 	cout << Form("XBS %10d %5d %10d  %4d  %5lu %6.1f    %8.4f %8.4f       %8.4f %8.4f      %8.4f %8.4f",
 		     run_,luminosityBlock_,event_,bunchCrossing_,
-      		   recVtxs->begin()->tracksSize(), recVtxs->begin()->ndof(),
+		     (unsigned long)(recVtxs->begin()->tracksSize()), recVtxs->begin()->ndof(),
       		   recVtxs->begin()->x(), 		   recVtxs->begin()->xError(), 
       		   recVtxs->begin()->y(), 		   recVtxs->begin()->yError(), 
       		   recVtxs->begin()->z(), 		   recVtxs->begin()->zError()
